+++ CHANGELOG.md
@@ -1,3 +1,21 @@
+# 48.14.0 (2022-11-28)
+
+
+### Features
+
+* **plasma-mantine:** export Header interface (#2953) 9545826, closes #2953
+
+
+
+## 48.13.2 (2022-11-28)
+
+
+### Bug Fixes
+
+* **plasma-mantine:** fix ModalWizard breaks when mantine version > 5.6.2 (#2952) ea04aed, closes #2952
+
+
+
 ## 48.13.1 (2022-11-24)